backend/requirements.txt
Django==4.1.6
seaborn~=0.12.2
pandas~=2.1.0
numpy~=1.25.2
scikit-learn~=1.3.0
transformers~=4.33.1
pyspellchecker~=0.7.2
sentence-transformers~=2.2.2
